Move heat flux code into flow module
